+++ src/App.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react'
 import './App.css'
 import request from 'superagent'
 import {} from 'dotenv/config'
+import Photo from './Photo'
 
 const token = process.env.REACT_APP_UNSPLASH_KEY
 
@@ -38,7 +39,6 @@ class App extends Component {
   }
 
   handleChange (e) {
-    e.preventDefault()
     this.setState({
       value: e.target.value
     })
@@ -68,61 +68,6 @@ class App extends Component {
         </div>
       </div>
     )
-  }
-}
-
-class Photo extends Component {
-  constructor () {
-    super()
-    this.state = {
-      isExpanded: false
-    }
-    this.expand = this.expand.bind(this)
-  }
-
-  expand () {
-    this.setState({
-      isExpanded: !this.state.isExpanded
-    })
-  }
-
-  render () {
-    let pic = this.props.pic
-    return (
-      <div>
-        <Modal isExpanded={this.state.isExpanded} pic={pic} expand={this.expand} />
-        <img className='thumb' src={pic.urls.thumb} alt='' onClick={this.expand} />
-      </div>
-    )
-  }
-}
-
-class Modal extends Component {
-  render () {
-    let isExpanded = this.props.isExpanded
-    let pic = this.props.pic
-    let expand = this.props.expand
-    if (isExpanded === true) {
-      return (
-        <div className='modal is-active'>
-          <div className='modal-background' />
-          <div className='modal-content'>
-            <p className='image is-4by3'>
-              <img src={pic.urls.regular} alt='' />
-            </p>
-            <div className='info'>
-              <h2 className='info__artist'>Artist: {pic.user.name}</h2>
-              <h2 className='info__portfolio'><a href={pic.user.portfolio_url} target='_blank'>View Portfolio</a></h2>
-              <h2 className='info__twitter'><a href={'https://www.twitter.com/' + pic.user.twitter_username} target='_blank'>View Twitter</a></h2>
-              <h2 className='info__instagram'><a href={'https://www.instagram.com/' + pic.user.instagram_username} target='_blank'>View Instagram</a></h2>
-            </div>
-          </div>
-          <button className='modal-close is-large' aria-label='close' onClick={expand} />
-        </div>
-      )
-    } else {
-      return null
-    }
   }
 }
